mm-video-legacy: venc: update metadata-type used for HAL1 based  recording
Update metadata buffer type from kMetadataBufferTypeCameraSource to
kMetadataBufferTypeNativeHandleSource as the latter is used by
camera-HALv1 that resides in a separate process.

Surface recording uses a different metadata as compared to
legacy camera metadata (HAL1).Cast the gralloc-meta to
appropriate struct to extract the handle

Since buffer-size is not known conclusively at the allocation
time of meta-handles, allow the least size to be allocated.
But check the buffer size before accessing in ETB

Bug: 26268807
Change-Id: I38e49d1e70f606470c48d603663451a1efec6360
Signed-off-by: AKabhishek <abhishekkaushik.star@gmail.com>